WIP: overhaul color changing to use only the paper object.
next step is fully ridding color changing of any dependency on the javascript object. currently, on importing an SVG, we construct a javascript object that represents the SVG's component tree. To change color, we traverse the javascript object, find the pointer to the paper element, and change the fill color that way. now, we instead directly query the element using paper's hitTest and change the fill color with the resulting pointer. the only challenge now is linking the color control bar and other DOM elements with the paper components.